# openapi.yaml
openapi: 3.1.0
info:
  title: Affilite Mix API
  version: "1"
  description: "Multi-tenant affiliate marketing platform API. Routes marked x-internal: true are not part of the public consumer contract."
  contact:
    name: Affilite Mix
  license:
    name: Proprietary
servers:
  - url: "https://compareai.site"
    description: Production (compareai.site)
  - url: "https://staging.compareai.site"
    description: Staging
paths:
...
  /api/admin/products/import:
    post:
      summary: POST /api/admin/products/import (site) — auth=admin
      operationId: postApiAdminProductsImport
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: ProductImportInput (multipart CSV)
      x-responseSchema: "{ imported: number; errors: ImportError[] }"
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: "Successful response ({ imported: number; errors: ImportError[] })"
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
...
  /api/admin/sites/select:
    post:
      summary: POST /api/admin/sites/select (site) — auth=admin
      operationId: postApiAdminSitesSelect
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: "{ siteId: string }"
      x-responseSchema: Ok
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: Successful response (Ok)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
...
components:
...
  headers:
    API-Version:
      description: Numeric major version of the served API contract.
      schema:
        type: string
        pattern: "^[1-9][0-9]*$"
        example: "1"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...